+++ src/main/java/es/codeurjc/board/rest/restController/ReviewRestController.java
@@ -10,9 +10,11 @@ import es.codeurjc.board.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Optional;
 import java.security.Principal;
 import java.util.List;
 
@@ -29,13 +31,22 @@ public class ReviewRestController {
     @Autowired
     private UserService userService;
 
-
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReview(@PathVariable long id) {
+    public ResponseEntity<ReviewDTO> deleteReview(@PathVariable long id, HttpServletRequest request) {
+        Review review = reviewsService.findById(id);
+        if (review != null && (reviewsService.reviewBelongsToUser(userService.getUser(request), reviewsService.findById(id)) || 
+            userService.isUserAdmin(request))) {
+            reviewsService.deleteById(id);
+            return ResponseEntity.ok(reviewMapper.toDTO(review));
+        }else{
+            if(!userService.seeIfUserIsLoggedIn(request)){
+                    return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); 
+                }
+                return ResponseEntity.notFound().build();
+        }
+        
 
-        reviewsService.deleteById(id);
 
-        return ResponseEntity.noContent().build();
     }
 
     @GetMapping
@@ -58,19 +69,24 @@ public class ReviewRestController {
     @PutMapping("/{id}")
     public ResponseEntity<ReviewDTO> updateReview(
             @PathVariable Long id,
-            @RequestBody ReviewDTO dto) {
-
+            @RequestBody ReviewDTO dto, HttpServletRequest request) {
+        
         Review review = reviewsService.findById(id);
+        
+        if(userService.seeIfUserIsLoggedIn(request) && !userService.isUserAdmin(request) &&reviewsService.reviewBelongsToUser(userService.getUser(request), review)){
+                        review.setTitle(dto.title());
+            review.setDescription(dto.description());
+            review.setType(dto.type());
 
-        review.setTitle(dto.title());
-        review.setDescription(dto.description());
-        review.setType(dto.type());
+            reviewsService.save(review, review.getUser());
 
-        reviewsService.save(review, review.getUser());
+            return ResponseEntity.ok(
+                    reviewMapper.toDTO(review)
+            );
+        }else{
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); 
+        }
 
-        return ResponseEntity.ok(
-                reviewMapper.toDTO(review)
-        );
     }
 
     @PostMapping
@@ -79,27 +95,30 @@ public class ReviewRestController {
             HttpServletRequest request) {
 
         User user = userService.getUser(request);
+        if(userService.seeIfUserIsLoggedIn(request)){
+                        Review review = new Review(
+                    dto.title(),
+                    dto.description(),
+                    dto.type()
+            );
 
-        Review review = new Review(
-                dto.title(),
-                dto.description(),
-                dto.type()
-        );
+            reviewsService.save(review, user);
+            boolean associationOk = true;  
+            if (dto.type() == Review.ReviewType.PLANT) {
+                associationOk = reviewsService.handlePlant(review,dto.plantSpecies());
+            } else if (dto.type() == Review.ReviewType.PRODUCT) {
+                associationOk = reviewsService.handleProduct(review, dto.product());
+            }
 
-        reviewsService.save(review, user);
+            if (!associationOk) {
+                return ResponseEntity.badRequest().build();
+            }
+                return ResponseEntity.status(201)
+                    .body(reviewMapper.toDTO(review));
+        }else{
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); 
+        }
 
-        return ResponseEntity.status(201)
-                .body(reviewMapper.toDTO(review));
     }
 
-
 }
-
-
-
-
-
-
-
-
-
